1. Added extensions To the error response.extensions type is Map<String, Any?>? 2. Added "error type" to  make it easy to branch by error types at the client side. 3. Added debug mode. If true, exception information will be output in extensions when an error occurs. 4. Embed serialize function to GraphQLError class because we want to use GraphQLError individually. ex: at StatusPages Plugin.
